desafiate finished

+++ js/login.js
@@ -46,6 +46,10 @@ function onSignIn(googleUser) {
 //   });
 
 function signOut() {
-  localStorage.clear();
-  location.href = 'https://mail.google.com/mail/u/0/?logout&hl=en';
+  if (localStorage.getItem('username')) {
+    localStorage.clear();
+    location.href = 'login.html';
+  } else {
+    location.href = 'https://mail.google.com/mail/u/0/?logout&hl=en';
+  }
 }
